Add messagebus for module communication
This commit adds the messagebus node module that will be used for all
module communication. It has been used to replace the communication
between the terminal and the input checker.

To avoid conflicts caused by the shared global scope of input_checker
and the terminal, input_checker has been moved to the mainview. In the
future, input_checker will be contained in its own module.

[#9]