Implement gita command for performing dry runs
As part of this change, how the terminal blocks had to be slightly
modified. Specifically, its behaviour had to be modified such that it
could print the output of a command instead of static text. One
functional change did occur as a result. That is, the command the user
typed is no longer deleted upon the command be blocked, as desired.

[#29]